+++ zcash_client_sqlite/src/wallet/init/migrations/received_notes_nullable_nf.rs
@@ -232,7 +232,7 @@ mod tests {
 
     use crate::{
         testing::db::{test_clock, test_rng},
-        wallet::init::{init_wallet_db_internal, migrations::v_transactions_net},
+        wallet::init::{migrations::v_transactions_net, WalletMigrator},
         WalletDb,
     };
 
@@ -246,13 +246,10 @@ mod tests {
             test_rng(),
         )
         .unwrap();
-        init_wallet_db_internal(
-            &mut db_data,
-            None,
-            &[v_transactions_net::MIGRATION_ID],
-            false,
-        )
-        .unwrap();
+        WalletMigrator::new()
+            .ignore_seed_relevance()
+            .init_or_migrate_to(&mut db_data, &[v_transactions_net::MIGRATION_ID])
+            .unwrap();
 
         // Create an account in the wallet
         let usk0 = UnifiedSpendingKey::from_seed(&db_data.params, &[0u8; 32][..], AccountId::ZERO)
@@ -277,7 +274,10 @@ mod tests {
             VALUES (0, 3, 0, '', 5, '', 'nf_b', false);").unwrap();
 
         // Apply the current migration
-        init_wallet_db_internal(&mut db_data, None, &[super::MIGRATION_ID], false).unwrap();
+        WalletMigrator::new()
+            .ignore_seed_relevance()
+            .init_or_migrate_to(&mut db_data, &[super::MIGRATION_ID])
+            .unwrap();
 
         {
             let mut q = db_data